% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandomGetterMethods.R
\docType{methods}
\name{getRandom,REARandom,logical,logical,logical,logical-method}
\alias{getRandom,REARandom,logical,logical,logical,logical-method}
\title{getRandom}
\usage{

  \S4method{getRandom}{REARandom,logical,logical,logical,logical}(REARandomObject,
  simplifyVector = TRUE, simplifyDataFrame = FALSE,
  simplifyMatrix = FALSE, flatten = FALSE)
}
\arguments{
\item{REARandomObject}{an Object of class REARandom passed to Random.org API
to retrieve the specified type of random variates.}

\item{simplifyVector}{A logical passed to the underlying call to jsonlite::fromJSON()}

\item{simplifyDataFrame}{A logical passed to the underlying call to jsonlite::fromJSON()}

\item{simplifyMatrix}{A logical passed to the underlying call to jsonlite::fromJSON()}

\item{flatten}{A logical passed to the underlying call to jsonlite::fromJSON()}
}
\description{
Method used to create REARandom object that will be passed in API
call to Random.org's generateIntegers method.
}
\examples{
\dontrun{

# Create a new REARandom object to get a series of random integer values
myNewReaRandomObject <- reaRandom() \%>\%
					   setIntegers(n = 1000,
					   			   min = 1000000,
					   			   max = 9999999) \%>\%
					   getRandom(simplifyVector = FALSE,
								  	 simplifyDataFrame = FALSE,
								  	 simplifyMatrix = FALSE,
								  	 flatten = FALSE)

}
}

