% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandom.R
\docType{methods}
\name{reaRandom}
\alias{reaRandom}
\title{reaRandom}
\usage{
reaRandom(api = NULL)
}
\arguments{
\item{api}{Takes a single character argument with either a file containing
the api key for Random.org, the key itself, or a NULL value.}
}
\value{
Returns an object of class REARandom
}
\description{
Function used to create new REARandom class objects
}
\details{
Function to construct new REARandom class objects
}
\examples{
\dontrun{

		# Create a new REARandom object in x with made up API key passed as
		# an argument to the function
		x <- reaRandom(api = "ABCD123456789abcd")

		# Same as above but assuming the API key is stored in a file on the
		# local system
		x <- reaRandom(api = "~/.randomOrgApiKey")

		# Same as above assuming the API key is stored as an option named "reaRandomKey"
		x <- reaRandom()

}
}

