% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandom.R
\docType{methods}
\name{setApiKey}
\alias{reaRandom-method}
\alias{setApiKey}
\alias{setApiKey,}
\alias{setApiKey,character-method}
\alias{setApiKey,missing-method}
\title{setApiKey}
\usage{
setApiKey(api)

\S4method{setApiKey}{character}(api = "character")

\S4method{setApiKey}{missing}(api = NULL)
}
\arguments{
\item{api}{A character string with the API key, a file name containing the API key, or NULL value}
}
\description{
Method used to set the ApiKey member of the REARandom class

Method used to allow user to pass a file name/string with the API key
}
\details{
Generic method declaration for API Key setter
}
\seealso{
Other REARandom.Constructor.Methods: \code{\link{setBLOBs,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,numeric,numeric-method}}

Other REARandom.Constructor.Methods: \code{\link{setBLOBs,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,numeric,numeric-method}}

Other REARandom.Constructor.Methods: \code{\link{setBLOBs,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,numeric,numeric-method}}
}

