% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandom.R
\docType{methods}
\name{setApiKey}
\alias{setApiKey}
\alias{setApiKey,REARandom,character-method}
\alias{setApiKey,REARandom,missing-method}
\title{setApiKey}
\usage{
setApiKey(REARandomObject, api)

\S4method{setApiKey}{REARandom,character}(REARandomObject, api = "character")

\S4method{setApiKey}{REARandom,missing}(REARandomObject, api = NULL)
}
\arguments{
\item{REARandomObject}{an Object of class REARandom}

\item{api}{A character string with the API key, a file name containing the API key, or NULL value}
}
\description{
Method used to set the ApiKey member of the REARandom class

Method used to allow user to pass a file name/string with the API key
}
\details{
Generic method declaration for API Key setter
}
\seealso{
Other REARandom Constructor Methods: \code{\link{setBLOBs,REARandom,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,REARandom,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,REARandom,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,REARandom,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,REARandom,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,REARandom,numeric,numeric-method}}

Other REARandom Constructor Methods: \code{\link{setBLOBs,REARandom,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,REARandom,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,REARandom,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,REARandom,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,REARandom,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,REARandom,numeric,numeric-method}}

Other REARandom Constructor Methods: \code{\link{setBLOBs,REARandom,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,REARandom,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,REARandom,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,REARandom,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,REARandom,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,REARandom,numeric,numeric-method}}
}

