% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandomConstructorMethods.R
\docType{methods}
\name{setBLOBs,numeric,numeric,numeric,character-method}
\alias{setBLOBs,numeric,numeric,numeric,character-method}
\title{setBLOBs}
\usage{
\S4method{setBLOBs}{numeric,numeric,numeric,character}(id = 42, n = 1,
  size = 8, format = "base64")
}
\arguments{
\item{id}{An ID number used to match response to request. Defaults to 42.}

\item{n}{The number of random values to return must be in [1, 100]}

\item{size}{The size in bits of the objects to return (must be divisible by 8)}

\item{format}{Either 'base64' or 'hex' to specify how the BLOBs are serialized}
}
\description{
Method used to create REARandom object that will be passed in API
call to Random.org's generateBlobs method.
}
\examples{
\dontrun{

		# Create a new REARandom object
		myNewReaRandomObject <- reaRandom()

		# Set object for retrieval of Binary Large OBjects that are 4MB in Size
		# and serialized as Hexadecimal values
		myNewReaRandomObject.setBLOBs(id = 37, n = 12, size = 4096, format = "hex")

		# Same as above, but 0.5MB objects serialized as base64
		myNewReaRandomObject.setBLOBs(id = 37, n = 12, size = 512, format = "base64")

}
}
\seealso{
Other REARandom Constructor Methods: \code{\link{setApiKey}},
  \code{\link{setDecimals,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,numeric,numeric-method}}
}

