% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandomConstructorMethods.R
\docType{methods}
\name{setDecimals,numeric,numeric,numeric,logical-method}
\alias{setDecimals,numeric,numeric,numeric,logical-method}
\title{setDecimals}
\usage{
\S4method{setDecimals}{numeric,numeric,numeric,logical}(id = 42, n = 1,
  decimalPlaces = 1, replacement = FALSE)
}
\arguments{
\item{id}{An ID number used to match response to request. Defaults to 42.}

\item{n}{The number of random values to return. Must be in [1, 1e4]; defaults to 1.}

\item{decimalPlaces}{The number of decimal places to return.  Must be in [1, 20]; defaults to 1.}

\item{replacement}{Should numbers be sampled with or without replacement.
Default is FALSE (e.g., unique values).}
}
\description{
Method used to create REARandom object that will be passed in API
call to Random.org's generateDecimalFractions method.
}
\examples{
\dontrun{

# Create a new REARandom object
myNewReaRandomObject <- reaRandom()

# Set object for retrieval of fractional random variables
myNewReaRandomObject.setDecimals(id = 42,
								 	n = 300,
								 	decimalPlaces = 5,
								 	replacement = TRUE)

}
}
\seealso{
Other REARandom Constructor Methods: \code{\link{setApiKey}},
  \code{\link{setBLOBs,numeric,numeric,numeric,character-method}},
  \code{\link{setIntegers,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,numeric,numeric-method}}
}

