% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandomConstructorMethods.R
\docType{methods}
\name{setIntegers,REARandom,numeric,numeric,numeric,numeric,logical,numeric-method}
\alias{setIntegers,REARandom,numeric,numeric,numeric,numeric,logical,numeric-method}
\title{setIntegers}
\usage{

  \S4method{setIntegers}{REARandom,numeric,numeric,numeric,numeric,logical,numeric}(REARandomObject,
  id = 42, n = 1, min = 1, max = 999, replacement = FALSE,
  base = 10)
}
\arguments{
\item{REARandomObject}{an Object of class REARandom to prepare for
generateIntegers API call}

\item{id}{ID Used for retrieval and matching of API call request and payload; default is 42.}

\item{n}{Number of random values to generate; default is 1.}

\item{min}{The minimum value integer to use for the request; default is 1.}

\item{max}{The maximum value integer to use for the request; default is 999.}

\item{replacement}{Whether the values should be sampled with or without replacement
from the distribution.}

\item{base}{Which base value should the numbers be reported in;
defaults to base 10 (decimal).  A value of 2 would provide binary values,
a value of 8 would return octal values, and value of 16 would return
hexadecimal values.}
}
\description{
Method used to create REARandom object that will be passed in API
call to Random.org's generateIntegers method.
}
\examples{
\dontrun{

# Create a new REARandom object to request random Integer values
myNewReaRandomObject <- reaRandom() \%>\%
						   setIntegers(id = 12,
									   n = 1000,
									   min = 1000,
									   max = 9999,
									   replacement = TRUE,
									   base = 8)

}
}
\seealso{
Other REARandom Constructor Methods: \code{\link{setApiKey}},
  \code{\link{setBLOBs,REARandom,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,REARandom,numeric,numeric,numeric,logical-method}},
  \code{\link{setNormal,REARandom,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,REARandom,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,REARandom,numeric,numeric-method}}
}

