% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandomConstructorMethods.R
\docType{methods}
\name{setNormal,numeric,numeric,numeric,numeric,numeric-method}
\alias{setNormal,numeric,numeric,numeric,numeric,numeric-method}
\title{setNormal}
\usage{
\S4method{setNormal}{numeric,numeric,numeric,numeric,numeric}(id = 42,
  n = 1, mean = 0, standardDeviation = 1, significantDigits = 2)
}
\arguments{
\item{id}{An ID number used to match response to request. Defaults to 42.}

\item{n}{The number of random values to return. Must be in [1, 1e4]; defaults to 1.}

\item{mean}{Value of the mean for the distribution from which to draw random numbers.
Must be in [-1e6, 1e6]; defaults to 0.}

\item{standardDeviation}{Value for the standard deviation of the distribution.
Must be in [-1e6, 1e6]; defaults to 1.}

\item{significantDigits}{The number of digits beyond the decimal to retain.
Must be in [2, 20]; defaults to 2.}
}
\description{
Method used to create REARandom object that will be passed in API
call to Random.org's generateGaussians method.
}
\examples{
\dontrun{

		# Create a new REARandom object
		myNewReaRandomObject <- reaRandom()

		# Set object for retrieval of Gaussian Random variables with $\\mu$ = 100
		# \\& $\\sigma$ = 15.
		myNewReaRandomObject.setNormal(id = 37, mean = 100,
		standardDeviation = 15, significantDigits = 10)

}
}
\seealso{
Other REARandom Constructor Methods: \code{\link{setApiKey}},
  \code{\link{setBLOBs,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setStrings,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,numeric,numeric-method}}
}

