% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandomConstructorMethods.R
\docType{methods}
\name{setStrings,REARandom,numeric,numeric,numeric,character,logical-method}
\alias{setStrings,REARandom,numeric,numeric,numeric,character,logical-method}
\title{setStrings}
\usage{

  \S4method{setStrings}{REARandom,numeric,numeric,numeric,character,logical}(REARandomObject,
  id = 42, n = 1, length = 1, characters = c(letters, LETTERS, c(0:9),
  "!", "@", "#", "$", "\%", "^", "&", "-", ",", "?", "|", ":", ";", "<", ">",
  ",", ".", "_"), replacement = FALSE)
}
\arguments{
\item{REARandomObject}{an Object of class REARandom to prepare for
generateStrings API call}

\item{id}{An ID number used to match response to request. Defaults to 42.}

\item{n}{The number of random values to return must be in [1, 1e4]}

\item{length}{The number of characters to include in each random string (must be in [1, 20])}

\item{characters}{The set of characters from which to generate random strings.
Must have length <= 80.}

\item{replacement}{Should numbers be sampled with or without replacement.
Default is FALSE (e.g., unique values).}
}
\description{
Method used to create REARandom object that will be passed in API
call to Random.org's generateStrings method.
}
\examples{
\dontrun{

# Create a new REARandom object
myNewReaRandomObject <- reaRandom() \%>\%
						   setStrings(id = 37,
						   			  n = 12,
						   			  characters = myChars,
						   			  replacement = FALSE)

# Create object with characters to use in method call
myChars <- c(letters, LETTERS, c(0:9), "!", "@", "#", "$", "\%", "^",
					"&", "-", ",", "?", "|", ":", ";", "<", ">", ",", ".", "_")

}
}
\seealso{
Other REARandom Constructor Methods: \code{\link{setApiKey}},
  \code{\link{setBLOBs,REARandom,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,REARandom,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,REARandom,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,REARandom,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setUniqueID,REARandom,numeric,numeric-method}}
}

