% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandomConstructorMethods.R
\docType{methods}
\name{setUniqueID,REARandom,numeric,numeric-method}
\alias{setUniqueID,REARandom,numeric,numeric-method}
\title{setUniqueID}
\usage{
\S4method{setUniqueID}{REARandom,numeric,numeric}(REARandomObject, id = 42,
  n = 1)
}
\arguments{
\item{REARandomObject}{an Object of class REARandom to prepare for
generateUUIDs API call}

\item{id}{An ID number used to match response to request. Defaults to 42.}

\item{n}{The number of Unique Uniform IDentifiers to return (must be in [1, 1e3])}
}
\description{
Method used to create REARandom object that will be passed in API
call to Random.org's generateUUIDs method.
}
\examples{
\dontrun{

# Create a new REARandom object
myNewReaRandomObject <- reaRandom() \%>\%
					   setUniqueID(id = 53,
					   			   n = 500)

}
}
\seealso{
Other REARandom Constructor Methods: \code{\link{setApiKey}},
  \code{\link{setBLOBs,REARandom,numeric,numeric,numeric,character-method}},
  \code{\link{setDecimals,REARandom,numeric,numeric,numeric,logical-method}},
  \code{\link{setIntegers,REARandom,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,REARandom,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,REARandom,numeric,numeric,numeric,character,logical-method}}
}

