% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandom.R
\name{validREARandom}
\alias{validREARandom}
\title{Function used to validate REARandom objects}
\usage{
validREARandom(object)
}
\arguments{
\item{object}{The S4 class object being validated.}
}
\description{
Validation method for REARandom construction
}
\details{
Validation function for Class REARandom
}

