% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandomGetterMethods.R
\docType{methods}
\name{getRandom,logical,logical,logical,logical-method}
\alias{getRandom,logical,logical,logical,logical-method}
\title{getRandom}
\usage{
\S4method{getRandom}{logical,logical,logical,logical}(simplifyVector = TRUE,
  simplifyDataFrame = FALSE, simplifyMatrix = FALSE, flatten = FALSE)
}
\arguments{
\item{simplifyVector}{A logical passed to the underlying call to jsonlite::fromJSON()}

\item{simplifyDataFrame}{A logical passed to the underlying call to jsonlite::fromJSON()}

\item{simplifyMatrix}{A logical passed to the underlying call to jsonlite::fromJSON()}

\item{flatten}{A logical passed to the underlying call to jsonlite::fromJSON()}
}
\description{
Method used to create REARandom object that will be passed in API
call to Random.org's generateIntegers method.
}
\examples{
\dontrun{

# Create a new REARandom object
myNewReaRandomObject <- reaRandom()

# Set object for retrieval of random integer values
myNewReaRandomObject.setIntegers(n = 1000, min = 1000000, max = 9999999)

# Get the resulting request from Random.org
myNewReaRandomObject.getRandom(simplifyVector = FALSE,
								  simplifyDataFrame = FALSE,
								  simplifyMatrix = FALSE,
								  flatten = FALSE)

}
}

