% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaRandomConstructorMethods.R
\docType{methods}
\name{setDecimals,REARandom,numeric,numeric,numeric,logical-method}
\alias{setDecimals,REARandom,numeric,numeric,numeric,logical-method}
\title{setDecimals}
\usage{

  \S4method{setDecimals}{REARandom,numeric,numeric,numeric,logical}(REARandomObject,
  id = 42, n = 1, decimalPlaces = 1, replacement = FALSE)
}
\arguments{
\item{REARandomObject}{an Object of class REARandom to prepare for
generateDecimalFractions API call}

\item{id}{An ID number used to match response to request. Defaults to 42.}

\item{n}{The number of random values to return. Must be in [1, 1e4]; defaults to 1.}

\item{decimalPlaces}{The number of decimal places to return.  Must be in [1, 20]; defaults to 1.}

\item{replacement}{Should numbers be sampled with or without replacement.
Default is FALSE (e.g., unique values).}
}
\description{
Method used to create REARandom object that will be passed in API
call to Random.org's generateDecimalFractions method.
}
\examples{
\dontrun{

# Create a new REARandom object
myNewReaRandomObject <- reaRandom() \%>\%
					   setDecimals(id = 42,
								 	   n = 300,
								 	   decimalPlaces = 5,
								 	   replacement = TRUE)

}
}
\seealso{
Other REARandom Constructor Methods: \code{\link{setApiKey}},
  \code{\link{setBLOBs,REARandom,numeric,numeric,numeric,character-method}},
  \code{\link{setIntegers,REARandom,numeric,numeric,numeric,numeric,logical,numeric-method}},
  \code{\link{setNormal,REARandom,numeric,numeric,numeric,numeric,numeric-method}},
  \code{\link{setStrings,REARandom,numeric,numeric,numeric,character,logical-method}},
  \code{\link{setUniqueID,REARandom,numeric,numeric-method}}
}

